requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-colorfield==0.3.0
django-storages==1.9.1
djangorestframework==3.11.0
dropbox==10.1.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pylint==2.5.2
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
